stagesense/backend/requirements.txt
google-adk>=1.0.0
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=14.0
sse-starlette>=2.1.0
python-dotenv>=1.0.0
pydantic>=2.11.0
httpx>=0.28.0
